requirements.txt
charset-normalizer==3.1.0
Django==2.2.1
Faker==17.6.0
idna==3.4
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.30.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==2.0.2
wincertstore==0.2
